requirements.txt
argparse==1.2.1
gearman==2.0.2
ipython==1.1.0
lockfile==0.9.1
python-vagrant==0.4.0
wsgiref==0.1.2
